Move shared network utilities to a header
