+++ tests/src/NetworkRequest.cpp
@@ -2,6 +2,7 @@
 
 #include "common/network/NetworkManager.hpp"
 #include "common/network/NetworkResult.hpp"
+#include "NetworkHelpers.hpp"
 #include "Test.hpp"
 
 #include <QCoreApplication>
@@ -9,14 +10,6 @@
 using namespace chatterino;
 
 namespace {
-
-#ifdef CHATTERINO_TEST_USE_PUBLIC_HTTPBIN
-// Not using httpbin.org, since it can be really slow and cause timeouts.
-// postman-echo has the same API.
-const char *const HTTPBIN_BASE_URL = "https://postman-echo.com";
-#else
-const char *const HTTPBIN_BASE_URL = "http://127.0.0.1:9051";
-#endif
 
 QString getStatusURL(int code)
 {
@@ -27,46 +20,6 @@ QString getDelayURL(int delay)
 {
     return QString("%1/delay/%2").arg(HTTPBIN_BASE_URL).arg(delay);
 }
-
-class RequestWaiter
-{
-public:
-    void requestDone()
-    {
-        {
-            std::unique_lock lck(this->mutex_);
-            ASSERT_FALSE(this->requestDone_);
-            this->requestDone_ = true;
-        }
-        this->condition_.notify_one();
-    }
-
-    void waitForRequest()
-    {
-        using namespace std::chrono_literals;
-
-        while (true)
-        {
-            {
-                std::unique_lock lck(this->mutex_);
-                bool done = this->condition_.wait_for(lck, 10ms, [this] {
-                    return this->requestDone_;
-                });
-                if (done)
-                {
-                    break;
-                }
-            }
-            QCoreApplication::processEvents(QEventLoop::AllEvents);
-            QCoreApplication::sendPostedEvents(nullptr, QEvent::DeferredDelete);
-        }
-    }
-
-private:
-    std::mutex mutex_;
-    std::condition_variable condition_;
-    bool requestDone_ = false;
-};
 
 }  // namespace
